Agrego Productos Vendidos con Linq

+++ Controllers/DTOS/PostVenta.cs
@@ -2,11 +2,11 @@
 {
     public class PostVenta
     {
-        public int Id { get; set; }
+        public long Id { get; set; }
         public string Descripciones { get; set; }
         public decimal Costo { get; set; }
         public decimal PrecioVenta { get; set; }
         public int Stock { get; set; }
-        public int IdUsuario { get; set; }
+        public long IdUsuario { get; set; }
     }
 }
