+++ Controllers/DTOS/PutProducto.cs
@@ -2,11 +2,11 @@
 {
     public class PutProducto
     {
-        public int Id { get; set; }
+        public long Id { get; set; }
         public string Descripciones { get; set; }
         public double Costo { get; set; }
         public double PrecioVenta { get; set; }
         public int Stock { get; set; }
-        public int IdUsuario { get; set; }
+        public long IdUsuario { get; set; }
     }
 }
